+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Bugfix
+- Fixed erroneous tag filter.
+- Fixed missing filter for name in Producer and Consumer Binding queries.
+
 ### Removed
 - Removed postgres repository.